wip - more work on plugin infrastructure
