--- src/DiabloInterface.Test/Business/Services/GameServiceTest.cs
@@ -1,12 +1,0 @@
-using System;
-using System.Text;
-using System.Collections.Generic;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
-
-namespace DiabloInterface.Test.Business.Services
-{
-    [TestClass]
-    public class GameServiceTest
-    {
-    }
-}
