--- src/DiabloInterface.Test/packages.config
@@ -1,7 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Castle.Core" version="4.4.0" targetFramework="net452" />
-  <package id="Moq" version="4.13.1" targetFramework="net452" />
-  <package id="System.Runtime.CompilerServices.Unsafe" version="4.5.0" targetFramework="net452" requireReinstallation="true" />
-  <package id="System.Threading.Tasks.Extensions" version="4.5.1" targetFramework="net452" requireReinstallation="true" />
-</packages>
