+++ src/DiabloInterface.sln
@@ -5,8 +5,6 @@ VisualStudioVersion = 15.0.27130.2003
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DiabloInterface", "DiabloInterface\DiabloInterface.csproj", "{25BB3310-9BD7-4234-AF7D-9F7FB6870A20}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DiabloInterface.Test", "DiabloInterface.Test\DiabloInterface.Test.csproj", "{D9A9D63E-4A48-4429-9098-478AEC6304E9}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "D2Reader", "D2Reader\D2Reader.csproj", "{C5F081D7-FB83-4934-908F-53CE986AB08C}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DiabloInterface.Plugin.PipeServer", "DiabloInterface.Plugin.PipeServer\DiabloInterface.Plugin.PipeServer.csproj", "{59944B50-1A5E-4E02-B9BE-0F9B499770DF}"
@@ -24,6 +22,16 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DiabloInterface.Plugin.Auto
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DiabloInterface.Plugin.FileWriter", "DiabloInterface.Plugin.FileWriter\DiabloInterface.Plugin.FileWriter.csproj", "{C9BA823B-E8D6-4B46-BC89-7E9377BB1815}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DiabloInterface.Plugin.Autosplits.Test", "DiabloInterface.Plugin.Autosplits.Test\DiabloInterface.Plugin.Autosplits.Test.csproj", "{6EA6B43B-B3CF-427E-9B28-C8736496CB9E}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DiabloInterface.Plugin.FileWriter.Test", "DiabloInterface.Plugin.FileWriter.Test\DiabloInterface.Plugin.FileWriter.Test.csproj", "{30DA147B-6D77-406E-8C08-0137105D1AED}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DiabloInterface.Core.Test", "DiabloInterface.Core.Test\DiabloInterface.Core.Test.csproj", "{C35E86F3-1281-4E85-BD8B-3BCB5EE9B7B1}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DiabloInterface.Plugin.PipeServer.Test", "DiabloInterface.Plugin.PipeServer.Test\DiabloInterface.Plugin.PipeServer.Test.csproj", "{7095BFA3-857A-4491-879F-B7D863E327E1}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "D2Reader.Test", "D2Reader.Test\D2Reader.Test.csproj", "{9D7AB953-19E0-46F6-A400-E3420C1E8905}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -34,10 +42,6 @@ Global
 		{25BB3310-9BD7-4234-AF7D-9F7FB6870A20}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{25BB3310-9BD7-4234-AF7D-9F7FB6870A20}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{25BB3310-9BD7-4234-AF7D-9F7FB6870A20}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D9A9D63E-4A48-4429-9098-478AEC6304E9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D9A9D63E-4A48-4429-9098-478AEC6304E9}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D9A9D63E-4A48-4429-9098-478AEC6304E9}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D9A9D63E-4A48-4429-9098-478AEC6304E9}.Release|Any CPU.Build.0 = Release|Any CPU
 		{C5F081D7-FB83-4934-908F-53CE986AB08C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{C5F081D7-FB83-4934-908F-53CE986AB08C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C5F081D7-FB83-4934-908F-53CE986AB08C}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -62,6 +66,26 @@ Global
 		{C9BA823B-E8D6-4B46-BC89-7E9377BB1815}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C9BA823B-E8D6-4B46-BC89-7E9377BB1815}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{C9BA823B-E8D6-4B46-BC89-7E9377BB1815}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6EA6B43B-B3CF-427E-9B28-C8736496CB9E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6EA6B43B-B3CF-427E-9B28-C8736496CB9E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6EA6B43B-B3CF-427E-9B28-C8736496CB9E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6EA6B43B-B3CF-427E-9B28-C8736496CB9E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{30DA147B-6D77-406E-8C08-0137105D1AED}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{30DA147B-6D77-406E-8C08-0137105D1AED}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{30DA147B-6D77-406E-8C08-0137105D1AED}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{30DA147B-6D77-406E-8C08-0137105D1AED}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C35E86F3-1281-4E85-BD8B-3BCB5EE9B7B1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C35E86F3-1281-4E85-BD8B-3BCB5EE9B7B1}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C35E86F3-1281-4E85-BD8B-3BCB5EE9B7B1}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C35E86F3-1281-4E85-BD8B-3BCB5EE9B7B1}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7095BFA3-857A-4491-879F-B7D863E327E1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7095BFA3-857A-4491-879F-B7D863E327E1}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7095BFA3-857A-4491-879F-B7D863E327E1}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7095BFA3-857A-4491-879F-B7D863E327E1}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9D7AB953-19E0-46F6-A400-E3420C1E8905}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9D7AB953-19E0-46F6-A400-E3420C1E8905}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9D7AB953-19E0-46F6-A400-E3420C1E8905}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9D7AB953-19E0-46F6-A400-E3420C1E8905}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
