+++ src/DiabloInterface/App.config
@@ -46,4 +46,9 @@
             <appender-ref ref="RollingFile"/>
         </root>
     </log4net>
+    <runtime>
+        <assemblyBinding xmlns="urn:schemas-microsoft-com:asm.v1">
+            <probing privatePath="Plugins" />
+        </assemblyBinding>
+    </runtime>
 </configuration>
